# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Papel",
      "description": "Papel is a light-weight core banking application that powers banking operations like account creation, customer deposits, and withdrawals. This app is meant to support a single bank, where users can signup and create bank accounts online, but must visit the branch to withdraw or deposit money.",
      "contact": {}
    },
    "host": "localhost:3000",
    
    "securityDefinitions": {},
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    
    "paths": {
      "/api/v1/signup": {
        "post": {
          "description": "",
          "summary": "Sign up",
          "tags": [
            "Sign"
          ],
          "operationId": "signUp", 
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            
            {
                "name": "username",
                "in": "formData",
                "required": true,
                "type": "string"
              },
              {
                "name": "firstname",
                "in": "formData",
                "required": true,
                "type": "string"
                
              },
              {
                "name": "lastname",
                "in": "formData",
                "required": true,
                "type": "string"
                
              },
            {
              "name": "email",
              "in": "formData",
              "required": true,
              "type": "string",
              "format": "email"
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "type": "string",
              "format":"password"
            }
          ],
          "responses": {
            "201": {
              "description": "User created successfully"
            },
           "400": {
              "description": "The user exists"
            }
          }
        }              
      },

      "/api/v1/signin": {
        "post": {
          "description": "",
          "summary": "user Sign in",
          "tags": [
            "Sign"
          ],
          "operationId": "signIn",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
          
            {
              "name": "email",
              "in": "formData",
              "required": true,
              "type": "string",
              "format": "email"
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "type": "string",
              "format": "password"
            }
          ],
          "responses": {
            "200": {
              "description": "User signed in successfully"
            },
            "400": {
              "description": "Incorrect password"
            },
            "404": {
              "description": "User does not exist!"
            }
          }
        }
      },
      
      "/api/v1/accounts": {
        "post": {
          "description": "Creating bank account to start",
          "summary": "User create account",
          "tags": [
            "User"
          ],
          "operationId": "createAccount",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "type": "string",
              "required":true
            },
            {
              "name": "accountType",
              "in": "formData",
              "default": "Savings",
              "enum": [
                "Loan",
                "Current",
                "Savings"
              ],
              "required": true,
              "type": "string"
            }
          ],

          "responses": {
            "200": {
              "description": "An account created successfully"
            },
            "400": {
              "description": "Try again with valid data"
            },
            "401": {
              "description": "Unauthorised: Sign in or Create account first!"
            }
          }
        },

        "get": {
          "description": "",
          "summary": "Cashier/Admin - View a list of all bank accounts​",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAccounts",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Accounts retrieved successfully!"
            },
            "400": {
              "description": "No account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },

      "/api/v1/accounts/{accNumber}": {
        "get": {
          "description": "",
          "summary": "User - View a specific bank account’s details",
          "tags": [
            "User"
          ],
          "operationId": "getOneAccount",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "An account retrieved successfully!"
            },
            "404": {
              "description": "No such account found"
            }
            
          }
        },

        "delete": {
          "description": "",
          "summary": "Cashier/Admin - Delete a specific bank account",
          "tags": [
            "Cashier"
          ],
          "operationId": "deleteOneAccount",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Account deleted successfully!"
            },
            "400": {
              "description": "No such account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        },

        "patch": {
          "description": "",
          "summary": "Admin - Activate or deactivate an account.",
          "tags": [
            "Administrator"
          ],
          "operationId": "toggleAccountStatus",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Status updated successfully"
            },
            "404": {
              "description": "No such account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin as Admin to proceed"
            }
            
          }
        }

      },

      "/api/v1/accounts/{accNumber}/transactions": {
        "get": {
          "description": "",
          "summary": "User - View an account’s transaction history​",
          "tags": [
            "User"
          ],
          "operationId": "getOneAccountTransactions",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "display bank account transactions"
            },
            "400": {
              "description": "No such transactions found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },



      "/api/v1/users": {
        "get": {
          "description": "",
          "summary": "Cashier or Admin - Get all users",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAllUsers",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Returns all Users"
            },
            "404": {
              "description": "No User found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },

      "/api/v1/users/{emailId}/accounts": {
        "get": {
          "description": "",
          "summary": "Cashier - View all accounts owned by a specific user (client)",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAccByUser",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "emailId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "display all accounts of a given user"
            },
            "400": {
              "description": "No User found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },

      "/api/v1/accounts?status=Active": {
        "get": {
          "description": "",
          "summary": "Cashier/Admin - View all active bank accounts",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAccounts",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Accounts retrieved successfully"
            },
            "400": {
              "description": "No account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },

      "/api/v1/accounts?status=Dormant": {
        "get": {
          "description": "",
          "summary": "Cashier/Admin - View all dormant bank accounts",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAccounts",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Accounts successfully retrieved!"
            },
            "400": {
              "description": "No account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },


      "/api/v1/transactions": {
        "get": {
          "description": "",
          "summary": "Cashier - Get all transactions",
          "tags": [
            "Cashier"
          ],
          "operationId": "getAllTransactions",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "display specified user account"
            },
            "400": {
              "description": "account number not found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },


      "/api/v1/transactions/3": {
        "get": {
          "description": "",
          "summary": "Cashier - View a specific transaction​",
          "tags": [
            "Cashier"
          ],
          "operationId": "getOneTransaction",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "display specified user account"
            },
            "400": {
              "description": "account number not found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },

      "/api/v1/transactions/{accNumber}/credit": {
        "post": {
          "description": "",
          "summary": "Cashier - Credit a bank account.",
          "tags": [
            "Cashier"
          ],
          "operationId": "creditAccount",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "amount",
              "in": "formData",
              "required": true,
              "type": "number"
            }
          ],
          "responses": {
            "200": {
              "description": "An account credit successfully"
            },
            "400": {
              "description": "No such account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },


      "/api/v1/transactions/{accNumber}/debit": {
        "post": {
          "description": "",
          "summary": "Cashier - Debit a bank account.",
          "tags": [
            "Cashier"
          ],
          "operationId": "debitAccount",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "accNumber",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "amount",
              "in": "formData",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "An account debit successfully"
            },
            "400": {
              "description": "No such account found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      },



      "/api/v1/users/{emailId}/switchrole": {
        "patch": {
          "description": "",
          "summary": "Admin - Switch user role.",
          "tags": [
            "Administrator"
          ],
          "operationId": "toggleAccounttype",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
            },
            {
              "name": "emailId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "display specified user account"
            },
            "400": {
              "description": "account number not found"
            },
            "401": {
              "description": "Unauthorised: Authorization not supplied, You must signin to proceed"
            }
            
          }
        }
      }











}
}